+++ src/main/java/de/derivo/neo2rdf/conversion/Neo4jToRDFConverter.java
@@ -5,9 +5,7 @@ import de.derivo.neo2rdf.conversion.config.ConversionConfig;
 import de.derivo.neo2rdf.conversion.model.Neo4jToRDFMapper;
 import de.derivo.neo2rdf.conversion.model.Neo4jToRDFMapperBuilder;
 import de.derivo.neo2rdf.conversion.model.Neo4jToRDFValueFactory;
-import de.derivo.neo2rdf.schema.IndexedNeo4jSchema;
-import de.derivo.neo2rdf.schema.IndexedNeo4jSchemaGenerator;
-import de.derivo.neo2rdf.util.ConsoleUtil;
+import de.derivo.neo2rdf.processors.Neo4jDBConnector;
 import org.eclipse.collections.impl.set.mutable.UnifiedSet;
 import org.eclipse.rdf4j.model.Resource;
 import org.eclipse.rdf4j.model.Statement;
@@ -15,66 +13,46 @@ import org.eclipse.rdf4j.model.ValueFactory;
 import org.eclipse.rdf4j.model.vocabulary.OWL;
 import org.eclipse.rdf4j.model.vocabulary.RDF;
 import org.eclipse.rdf4j.model.vocabulary.RDFS;
-import org.neo4j.kernel.impl.store.NeoStores;
 import org.roaringbitmap.longlong.Roaring64Bitmap;
-import org.slf4j.Logger;
+import org.tinylog.Logger;
 
 import java.util.Map;
 import java.util.Set;
 
 public abstract class Neo4jToRDFConverter {
 
-    protected Logger log = ConsoleUtil.getLogger();
-    private final NeoStores neoStores;
     private NodeToRDFConverter nodeProcessor;
     private RelationshipToRDFConverter relationshipProcessor;
-    private boolean includeDeletedNeo4jLabels = false;
-    private boolean includeDeletedPropertyKeys = false;
-    private boolean includeDeletedRelationshipTypes = false;
-
-    private IndexedNeo4jSchema indexedSchema;
 
     protected Neo4jToRDFMapper neo4jToRDFMapper;
+    protected IndexedNeo4jSchema indexedSchema;
 
-    private Map<Long, String> neo4jLabelIDToStr;
-    private Map<Long, String> propertyKeyIDToStr;
-    private Map<Long, String> relationshipTypeIDToStr;
-
-    private Set<Long> deployedNeo4jLabels;
-    private Set<Long> deployedRelationshipTypes;
-    private Set<Long> deployedPropertyKeys;
-    private Set<Long> datatypePropertyKeys;
-    private Set<Long> objectPropertyKeys;
-    private Set<Long> annotationPropertyKeys;
+    private Set<String> deployedNeo4jLabels;
+    private Set<String> deployedRelationshipTypes;
+    private Set<String> deployedPropertyKeys;
+    private Set<String> datatypePropertyKeys;
+    private Set<String> objectPropertyKeys;
+    private Set<String> annotationPropertyKeys;
 
     private final ValueFactory valueFactory = new Neo4jToRDFValueFactory();
 
+    private final Neo4jDBConnector neo4jDBConnector;
     protected final ConversionConfig config;
 
-    public Neo4jToRDFConverter(NeoStores neoStores, ConversionConfig config) {
-        this.neoStores = neoStores;
+    public Neo4jToRDFConverter(Neo4jDBConnector neo4jDBConnector, ConversionConfig config) {
+        this.neo4jDBConnector = neo4jDBConnector;
         this.config = config;
         init();
     }
 
     protected void init() {
-        this.includeDeletedRelationshipTypes = config.isIncludeDeletedRelationshipTypes();
-        this.includeDeletedNeo4jLabels = config.isIncludeDeletedNeo4jLabels();
-        this.includeDeletedPropertyKeys = config.isIncludeDeletedPropertyKeys();
-
-        log.info("Indexing schema of dataset...");
-        IndexedNeo4jSchemaGenerator indexedSchemaGenerator = new IndexedNeo4jSchemaGenerator(neoStores);
-        this.indexedSchema = indexedSchemaGenerator.generate();
-        this.neo4jLabelIDToStr = indexedSchema.getLabelIDToStr();
-        this.propertyKeyIDToStr = indexedSchema.getPropertyKeyIDToStr();
-        this.relationshipTypeIDToStr = indexedSchema.getRelationshipTypeIDToStr();
-
-        Neo4jToRDFMapperBuilder builder = new Neo4jToRDFMapperBuilder(config.getBasePrefix(), this.indexedSchema);
+        Neo4jToRDFMapperBuilder builder = new Neo4jToRDFMapperBuilder(config.getBasePrefix());
         builder.setReificationVocabulary(config.getReificationVocabulary());
         this.neo4jToRDFMapper = builder.build();
 
-        nodeProcessor = new NodeToRDFConverter(neoStores, this, config);
-        relationshipProcessor = new RelationshipToRDFConverter(neoStores, this, config);
+        indexedSchema = new IndexedNeo4jSchema(neo4jDBConnector);
+        nodeProcessor = new NodeToRDFConverter(neo4jDBConnector, this, config);
+        relationshipProcessor = new RelationshipToRDFConverter(neo4jDBConnector, this, config);
     }
 
 
@@ -84,13 +62,14 @@ public abstract class Neo4jToRDFConverter {
 
     public void startProcessing() {
         onStart();
-        log.info("Processing axiomatic Neo4j triples...");
+        Logger.info("Processing axiomatic Neo4j triples...");
         processNeo4jAxiomaticTriples();
-        log.info("Processing nodes...");
+        Logger.info("Processing nodes...");
         nodeProcessor.startProcessing();
-        this.deployedPropertyKeys = new UnifiedSet<>(indexedSchema.getPropertyKeyIDToStr().size());
-        this.objectPropertyKeys = new UnifiedSet<>(indexedSchema.getPropertyKeyIDToStr().size());
-        this.datatypePropertyKeys = new UnifiedSet<>(indexedSchema.getPropertyKeyIDToStr().size());
+        int defaultMaxExpectedNumberOfProperties = 1_000;
+        this.deployedPropertyKeys = new UnifiedSet<>(defaultMaxExpectedNumberOfProperties);
+        this.objectPropertyKeys = new UnifiedSet<>(defaultMaxExpectedNumberOfProperties);
+        this.datatypePropertyKeys = new UnifiedSet<>(defaultMaxExpectedNumberOfProperties);
 
         this.deployedNeo4jLabels = this.nodeProcessor.getDeployedNeo4jLabels();
         this.deployedPropertyKeys.addAll(this.nodeProcessor.getDatatypePropertyKeys());
@@ -99,10 +78,10 @@ public abstract class Neo4jToRDFConverter {
         this.datatypePropertyKeys.addAll(this.nodeProcessor.getDatatypePropertyKeys());
         this.objectPropertyKeys.addAll(this.nodeProcessor.getObjectPropertyKeys());
 
-        log.info("Processing relationships...");
+        Logger.info("Processing relationships...");
         relationshipProcessor.startProcessing();
         this.deployedRelationshipTypes = relationshipProcessor.getDeployedRelationshipTypes();
-        this.annotationPropertyKeys = new UnifiedSet<>(indexedSchema.getPropertyKeyIDToStr().size());
+        this.annotationPropertyKeys = new UnifiedSet<>(defaultMaxExpectedNumberOfProperties);
         this.annotationPropertyKeys.addAll(relationshipProcessor.getObjectPropertyKeys());
         this.annotationPropertyKeys.addAll(relationshipProcessor.getDatatypePropertyKeys());
         this.deployedPropertyKeys.addAll(this.annotationPropertyKeys);
@@ -111,19 +90,19 @@ public abstract class Neo4jToRDFConverter {
         this.objectPropertyKeys.addAll(this.relationshipProcessor.getObjectPropertyKeys());
 
         if (this.config.isDeriveClassHierarchyByLabelSubsetCheck()) {
-            log.info("Deriving class hierarchy based on Neo4j label subset check...");
-            deriveClassHierarchy(nodeProcessor.getLabelIDToInstanceSet());
+            Logger.info("Deriving class hierarchy based on Neo4j label subset check...");
+            deriveClassHierarchy(nodeProcessor.getLabelToInstanceSet());
         }
 
         if (this.config.isDerivePropertyHierarchyByRelationshipSubsetCheck()) {
-            log.info("Deriving property hierarchy based on Neo4j relationship subset check...");
+            Logger.info("Deriving property hierarchy based on Neo4j relationship subset check...");
             derivePropertyHierarchy(relationshipProcessor.getRelationshipIDToInstanceSet());
         }
 
-        log.info("Adding schema for present labels, property keys, and relationship types...");
+        Logger.info("Adding schema for present labels, property keys, and relationship types...");
         processPropertyKeysAndLabels();
         onFinish();
-        log.info("Neo4j database successfully processed.");
+        Logger.info("Neo4j database successfully processed.");
     }
 
     private void processNeo4jAxiomaticTriples() {
@@ -131,15 +110,12 @@ public abstract class Neo4jToRDFConverter {
     }
 
     private void processPropertyKeysAndLabels() {
-        neo4jLabelIDToStr.forEach((labelID, str) -> {
-            if (!includeDeletedNeo4jLabels && !deployedNeo4jLabels.contains(labelID)) {
-                return;
-            }
-            Resource rdfClass = neo4jToRDFMapper.labelIDToResource(labelID);
+        indexedSchema.getNeo4jLabels().forEach((neo4jLabel) -> {
+            Resource rdfClass = neo4jToRDFMapper.labelToResource(neo4jLabel);
             Statement s = valueFactory.createStatement(
                     rdfClass,
                     RDFS.LABEL,
-                    valueFactory.createLiteral(str)
+                    valueFactory.createLiteral(neo4jLabel)
             );
             processStatement(s);
             s = valueFactory.createStatement(
@@ -150,23 +126,16 @@ public abstract class Neo4jToRDFConverter {
             processStatement(s);
         });
 
-        propertyKeyIDToStr.forEach((propertyKeyID, str) -> {
-            if (!includeDeletedPropertyKeys && !deployedPropertyKeys.contains(propertyKeyID)) {
-                return;
-            }
-            Resource dataProperty = neo4jToRDFMapper.propertyKeyIDToResource(propertyKeyID);
-            if (str.startsWith("__org.neo4j.SchemaRule.")) {
-                // skip schema rules
-                return;
-            }
+        indexedSchema.getPropertyKeys().forEach((propertyKey) -> {
+            Resource dataProperty = neo4jToRDFMapper.propertyKeyToResource(propertyKey);
             Statement s = valueFactory.createStatement(
                     dataProperty,
                     RDFS.LABEL,
-                    valueFactory.createLiteral(str)
+                    valueFactory.createLiteral(propertyKey)
             );
             processStatement(s);
 
-            if (datatypePropertyKeys.contains(propertyKeyID)) {
+            if (datatypePropertyKeys.contains(propertyKey)) {
                 s = valueFactory.createStatement(
                         dataProperty,
                         RDF.TYPE,
@@ -175,7 +144,7 @@ public abstract class Neo4jToRDFConverter {
                 processStatement(s);
             }
 
-            if (objectPropertyKeys.contains(propertyKeyID)) {
+            if (objectPropertyKeys.contains(propertyKey)) {
                 s = valueFactory.createStatement(
                         dataProperty,
                         RDF.TYPE,
@@ -184,7 +153,7 @@ public abstract class Neo4jToRDFConverter {
                 processStatement(s);
             }
 
-            if (annotationPropertyKeys.contains(propertyKeyID)) {
+            if (annotationPropertyKeys.contains(propertyKey)) {
                 s = valueFactory.createStatement(
                         dataProperty,
                         RDF.TYPE,
@@ -202,15 +171,12 @@ public abstract class Neo4jToRDFConverter {
 
         });
 
-        relationshipTypeIDToStr.forEach((relationshipTypeID, str) -> {
-            if (!includeDeletedRelationshipTypes && !deployedRelationshipTypes.contains(relationshipTypeID)) {
-                return;
-            }
-            Resource objectProperty = neo4jToRDFMapper.relationshipTypeIDToIRI(relationshipTypeID);
+        indexedSchema.getRelationshipTypes().forEach((relationshipType) -> {
+            Resource objectProperty = neo4jToRDFMapper.relationshipTypeToIRI(relationshipType);
             Statement s = valueFactory.createStatement(
                     objectProperty,
                     RDFS.LABEL,
-                    valueFactory.createLiteral(str)
+                    valueFactory.createLiteral(relationshipType)
             );
             processStatement(s);
             s = valueFactory.createStatement(
@@ -223,13 +189,13 @@ public abstract class Neo4jToRDFConverter {
 
     }
 
-    private void deriveClassHierarchy(Map<Long, Roaring64Bitmap> labelIDToInstanceSet) {
+    private void deriveClassHierarchy(Map<String, Roaring64Bitmap> labelIDToInstanceSet) {
         SubsetCheck subsetChecker = new SubsetCheck(labelIDToInstanceSet);
-        Map<Long, Set<Long>> subIDToSuperIDs = subsetChecker.deriveSubsumesRelations();
+        Map<String, Set<String>> subIDToSuperIDs = subsetChecker.deriveSubsumesRelations();
         subIDToSuperIDs.forEach((subID, superIDs) -> {
-            Resource subClassIRI = neo4jToRDFMapper.labelIDToResource(subID);
-            for (Long superID : superIDs) {
-                Resource superClassIRI = neo4jToRDFMapper.labelIDToResource(superID);
+            Resource subClassIRI = neo4jToRDFMapper.labelToResource(subID);
+            for (String superID : superIDs) {
+                Resource superClassIRI = neo4jToRDFMapper.labelToResource(superID);
                 Statement s = valueFactory.createStatement(
                         subClassIRI,
                         RDFS.SUBCLASSOF,
@@ -240,13 +206,13 @@ public abstract class Neo4jToRDFConverter {
         });
     }
 
-    private void derivePropertyHierarchy(Map<Long, Roaring64Bitmap> relationshipTypeIDToInstanceSet) {
+    private void derivePropertyHierarchy(Map<String, Roaring64Bitmap> relationshipTypeIDToInstanceSet) {
         SubsetCheck subsetChecker = new SubsetCheck(relationshipTypeIDToInstanceSet);
-        Map<Long, Set<Long>> subIDToSuperIDs = subsetChecker.deriveSubsumesRelations();
+        Map<String, Set<String>> subIDToSuperIDs = subsetChecker.deriveSubsumesRelations();
         subIDToSuperIDs.forEach((subID, superIDs) -> {
-            Resource subPropertyIRI = neo4jToRDFMapper.relationshipTypeIDToIRI(subID);
-            for (Long superID : superIDs) {
-                Resource superPropertyIRI = neo4jToRDFMapper.relationshipTypeIDToIRI(superID);
+            Resource subPropertyIRI = neo4jToRDFMapper.relationshipTypeToIRI(subID);
+            for (String superID : superIDs) {
+                Resource superPropertyIRI = neo4jToRDFMapper.relationshipTypeToIRI(superID);
                 Statement s = valueFactory.createStatement(
                         subPropertyIRI,
                         RDFS.SUBPROPERTYOF,