+++ src/main/java/de/derivo/neo2rdf/conversion/checks/SubsetCheck.java
@@ -5,26 +5,27 @@ import org.roaringbitmap.longlong.Roaring64Bitmap;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Set;
 
 public class SubsetCheck {
 
-    private final Map<Long, Roaring64Bitmap> idToInstancesSet;
-    private final Map<Long, Set<Long>> subsumesRelations;
+    private final Map<String, Roaring64Bitmap> idToInstancesSet;
+    private final Map<String, Set<String>> subsumesRelations;
 
-    public SubsetCheck(Map<Long, Roaring64Bitmap> idToInstancesSet) {
+    public SubsetCheck(Map<String, Roaring64Bitmap> idToInstancesSet) {
         this.idToInstancesSet = idToInstancesSet;
         this.subsumesRelations = new HashMap<>(idToInstancesSet.size());
     }
 
-    public Map<Long, Set<Long>> deriveSubsumesRelations() {
-        for (Map.Entry<Long, Roaring64Bitmap> subEntry : idToInstancesSet.entrySet()) {
-            long subConceptID = subEntry.getKey();
+    public Map<String, Set<String>> deriveSubsumesRelations() {
+        for (Map.Entry<String, Roaring64Bitmap> subEntry : idToInstancesSet.entrySet()) {
+            String subConceptID = subEntry.getKey();
             Roaring64Bitmap subInstances = subEntry.getValue();
 
-            for (Map.Entry<Long, Roaring64Bitmap> superEntry : idToInstancesSet.entrySet()) {
-                long superConceptID = superEntry.getKey();
-                if (subConceptID == superConceptID) {
+            for (Map.Entry<String, Roaring64Bitmap> superEntry : idToInstancesSet.entrySet()) {
+                String superConceptID = superEntry.getKey();
+                if (Objects.equals(subConceptID, superConceptID)) {
                     continue;
                 }
                 Roaring64Bitmap superInstances = superEntry.getValue();