+++ src/main/java/de/derivo/neo2rdf/conversion/cli/Neo4jToRDFConversionServerCommand.java
@@ -2,16 +2,20 @@ package de.derivo.neo2rdf.conversion.cli;
 
 import de.derivo.neo2rdf.conversion.Neo4jToTurtleConversionServer;
 import de.derivo.neo2rdf.conversion.config.ConversionConfig;
-import org.neo4j.kernel.impl.store.NeoStores;
 import picocli.CommandLine;
 
+@SuppressWarnings("CanBeFinal")
 @CommandLine.Command(name = "server",
         header = "Starts an HTTP server that serves the conversion result as RDF Turtle stream",
         showDefaultValues = true,
         usageHelpWidth = 95,
         description = """
                 The application starts an HTTP server for the provided Neo4j database. When a GET request is sent to the server, the conversion procedure is initiated and the response returns an RDF Turtle stream to the client.
-                Exemplary usage: `server -db=./path/to/neo4jdb -p=8080`
+                Exemplary usage: `server --database="neo4j" \\
+                                    --uri="neo4j+s://867928679.databases.neo4j.io" \\
+                                    --user="neo4j" \\
+                                    --password="eBWczH5dRt2VR1C1eYKvk5jRt2VR1C1eY72NUCk" \\
+                                    --port=8080`
                 """)
 public class Neo4jToRDFConversionServerCommand implements Runnable {
 
@@ -28,10 +32,8 @@ public class Neo4jToRDFConversionServerCommand implements Runnable {
 
     @Override
     public void run() {
-        NeoStores neoStores = options.getNeo4jStore();
         ConversionConfig config = options.getConversionConfig();
-        Neo4jToTurtleConversionServer server = new Neo4jToTurtleConversionServer(options.getNeo4jDBDirectory(),
-                neoStores,
+        Neo4jToTurtleConversionServer server = new Neo4jToTurtleConversionServer(options.getNeo4jDBConnector(),
                 config,
                 port, numberOfServerThreads);
         server.startServer();